requirements.txt
soco-cli
schedule
pytz